Simple stored procedure that returns a stored procedure version number
This stored proc will use a version number generated during the gradle build as the stored proc's version.  Helpful when identifying what stored procs are in a Neo4j environment.

This also includes the what is needed in the gradle build file to generate the version number.  There might be a better way to do this.

Author: Dan Dentinger